#8: restructure handlers, add contracts
